create comments and likes

+++ website/views.py
@@ -1,32 +1,34 @@
-from flask import Blueprint , render_template , redirect , flash , request
-from flask_login import login_required , current_user
-from .models import Post , User
+from flask import Blueprint, render_template, redirect, flash, request
+from flask_login import login_required, current_user
+from .models import Post, User
 from . import db
 
-views = Blueprint('views' , __name__)
+views = Blueprint('views', __name__)
 
 @views.route("/")
 @views.route("/home")
 @login_required
 def home():
     posts = Post.query.all()
-    return render_template("home.html" , posts=posts , user=current_user)
+    return render_template("home.html", posts=posts, user=current_user)
 
-@views.route("/create-post" , methods=['GET' , 'POST'])
+@views.route("/create-post", methods=['GET', 'POST'])
 @login_required
 def create_post():
     if request.method == 'POST':
         text = request.form.get('text')
         caption = request.form.get('caption')
         if len(text) < 2:
-            flash("Too short" , category='error')
+            flash("Too short", category='error')
         else:
-            new_post = Post(text=text , caption=caption ,author=current_user.id )
+            new_post = Post(text=text, caption=caption, author=current_user.id)
             db.session.add(new_post)
             db.session.commit()
-    return render_template("create_post.html")
+            return redirect("/")
+    return render_template("create_post.html", user=current_user)
 
-@views.route("/update/<post_id>" , methods=['GET' , 'POST'])
+@views.route("/update/<post_id>", methods=['GET', 'POST'])
+@login_required
 def update(post_id):
     post = Post.query.filter_by(id=post_id).first()
     if request.method == 'POST':
@@ -36,21 +38,23 @@ def update(post_id):
         post.text = text
         db.session.commit()
         return redirect("/")
-    return render_template("update_post.html",  caption=post.caption , text=post.text)
+    return render_template("update_post.html", caption=post.caption, text=post.text, user=current_user)
 
-@views.route("/delete/<post_id>" , methods=['GET' , 'POST'])
+@views.route("/delete/<post_id>", methods=['GET', 'POST'])
+@login_required
 def delete(post_id):
     post = Post.query.filter_by(id=post_id).first()
-    db.session.delete(post)
-    db.session.commit()
+    if post:
+        db.session.delete(post)
+        db.session.commit()
     return redirect("/")
 
-
-#last thing to do
 @views.route("/user/<name>")
 def user_profile(name):
     user = User.query.filter_by(username=name).first()
     if user:
-        return render_template("profile_page.html" , user=user)
+        return render_template("profile_page.html", user=user)
     else:
         return render_template("404.html")
+
+
